explicitly refuse dealing with symbolic links

+++ hatta.py
@@ -353,7 +353,8 @@ class WikiStorage(object):
 
     def __contains__(self, title):
         if title:
-            return os.path.exists(self._file_path(title))
+            file_path = self._file_path(title)
+            return os.path.isfile(file_path) and not os.path.islink(file_path)
 
     def __iter__(self):
         return self.all_pages()
@@ -392,6 +393,8 @@ class WikiStorage(object):
         text = comment.encode('utf-8') or _(u'comment').encode('utf-8')
         repo_file = self._title_to_file(title)
         file_path = self._file_path(title)
+        if os.path.islink(file_path):
+            raise werkzeug.exceptions.Forbidden(u"Can't edit symbolic links")
         mercurial.util.rename(file_name, file_path)
         changectx = self._changectx()
         try:
@@ -461,6 +464,8 @@ class WikiStorage(object):
         text = comment.encode('utf-8') or 'deleted'
         repo_file = self._title_to_file(title)
         file_path = self._file_path(title)
+        if os.path.islink(file_path):
+            raise werkzeug.exceptions.Forbidden(u"Can't edit symbolic links")
         try:
             os.unlink(file_path)
         except OSError:
@@ -471,8 +476,11 @@ class WikiStorage(object):
     def open_page(self, title):
         """Open the page and return a file-like object with its contents."""
 
+        file_path = self._file_path(title)
+        if os.path.islink(file_path):
+            raise werkzeug.exceptions.Forbidden(u"Can't read symbolic links")
         try:
-            return open(self._file_path(title), "rb")
+            return open(file_path, "rb")
         except IOError:
             raise werkzeug.exceptions.NotFound()
 
@@ -1883,6 +1891,9 @@ class WikiPageText(WikiPage):
         except werkzeug.exceptions.NotFound:
             comment = _(u'created')
             rev = -1
+        except werkzeug.exceptions.Forbidden:
+            yield werkzeug.html.p(werkzeug.html(_(u"Can't edit a symbolic link.")))
+            return
         if preview:
             lines = preview
             comment = self.request.form.get('comment', comment)
